fix min/max value for config: now double, not int fix some wrning

+++ src/libslic3r/Fill/Fill3DHoneycomb.cpp
@@ -154,8 +154,8 @@ void Fill3DHoneycomb::_fill_surface_single(
     Polylines   polylines = makeGrid(
         scale_(this->z),
         distance,
-        ceil(bb.size().x() / distance) + 1,
-        ceil(bb.size().y() / distance) + 1,
+        (size_t)ceil(bb.size().x() / distance) + 1,
+        (size_t)ceil(bb.size().y() / distance) + 1,
 		size_t((this->layer_id / thickness_layers) % 2) + 1);
 	//makeGrid(coord_t z, coord_t gridSize, size_t gridWidth, size_t gridHeight, size_t curveType)
